+++ Files/Files.csproj
@@ -211,6 +211,9 @@
     <Compile Include="Dialogs\ConfirmDeleteDialog.xaml.cs">
       <DependentUpon>ConfirmDeleteDialog.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Helpers\NavigationHelpers.cs" />
+    <Compile Include="Helpers\WallpaperHelpers.cs" />
+    <Compile Include="Helpers\Win32Helpers.cs" />
     <Compile Include="UserControls\RestartControl.xaml.cs">
       <DependentUpon>RestartControl.xaml</DependentUpon>
     </Compile>
