+++ Files/Interacts/Interaction.cs
@@ -112,108 +112,6 @@ namespace Files.Interacts
             }
         }
 
-        public async void SetAsBackground(WallpaperType type, string filePath)
-        {
-            if (UserProfilePersonalizationSettings.IsSupported())
-            {
-                // Get the path of the selected file
-                StorageFile sourceFile = await StorageItemHelpers.ToStorageItem<StorageFile>(filePath, AssociatedInstance);
-                if (sourceFile == null)
-                {
-                    return;
-                }
-
-                // Get the app's local folder to use as the destination folder.
-                StorageFolder localFolder = ApplicationData.Current.LocalFolder;
-
-                // the file to the destination folder.
-                // Generate unique name if the file already exists.
-                // If the file you are trying to set as the wallpaper has the same name as the current wallpaper,
-                // the system will ignore the request and no-op the operation
-                var file = (StorageFile)await FilesystemTasks.Wrap(() => sourceFile.CopyAsync(localFolder, sourceFile.Name, NameCollisionOption.GenerateUniqueName).AsTask());
-                if (file == null)
-                {
-                    return;
-                }
-
-                UserProfilePersonalizationSettings profileSettings = UserProfilePersonalizationSettings.Current;
-                if (type == WallpaperType.Desktop)
-                {
-                    // Set the desktop background
-                    await profileSettings.TrySetWallpaperImageAsync(file);
-                }
-                else if (type == WallpaperType.LockScreen)
-                {
-                    // Set the lockscreen background
-                    await profileSettings.TrySetLockScreenImageAsync(file);
-                }
-            }
-        }
-
-        public static async void OpenPathInNewTab(string path)
-        {
-            await MainPage.AddNewTabByPathAsync(typeof(PaneHolderPage), path);
-        }
-
-        public static async Task<bool> OpenPathInNewWindowAsync(string path)
-        {
-            var folderUri = new Uri($"files-uwp:?folder={Uri.EscapeDataString(path)}");
-            return await Launcher.LaunchUriAsync(folderUri);
-        }
-
-        public static async Task<bool> OpenTabInNewWindowAsync(string tabArgs)
-        {
-            var folderUri = new Uri($"files-uwp:?tab={Uri.EscapeDataString(tabArgs)}");
-            return await Launcher.LaunchUriAsync(folderUri);
-        }
-
-        public async void OpenDirectoryInTerminal(string workingDir)
-        {
-            var terminal = AppSettings.TerminalController.Model.GetDefaultTerminal();
-
-            if (Connection != null)
-            {
-                var value = new ValueSet
-                {
-                    { "WorkingDirectory", workingDir },
-                    { "Application", terminal.Path },
-                    { "Arguments", string.Format(terminal.Arguments,
-                       Helpers.PathNormalization.NormalizePath(workingDir)) }
-                };
-                await Connection.SendMessageAsync(value);
-            }
-        }
-
-        public async Task InvokeWin32ComponentAsync(string applicationPath, string arguments = null, bool runAsAdmin = false, string workingDir = null)
-        {
-            await InvokeWin32ComponentsAsync(new List<string>() { applicationPath }, arguments, runAsAdmin, workingDir);
-        }
-
-        public async Task InvokeWin32ComponentsAsync(List<string> applicationPaths, string arguments = null, bool runAsAdmin = false, string workingDir = null)
-        {
-            Debug.WriteLine("Launching EXE in FullTrustProcess");
-            if (Connection != null)
-            {
-                var value = new ValueSet
-                {
-                    { "WorkingDirectory", string.IsNullOrEmpty(workingDir) ? AssociatedInstance?.FilesystemViewModel?.WorkingDirectory : workingDir },
-                    { "Application", applicationPaths.FirstOrDefault() },
-                    { "ApplicationList", JsonConvert.SerializeObject(applicationPaths) },
-                };
-
-                if (runAsAdmin)
-                {
-                    value.Add("Arguments", "runas");
-                }
-                else
-                {
-                    value.Add("Arguments", arguments);
-                }
-
-                await Connection.SendMessageAsync(value);
-            }
-        }
-
         public async Task OpenShellCommandInExplorerAsync(string shellCommand)
         {
             Debug.WriteLine("Launching shell command in FullTrustProcess");
@@ -235,290 +133,6 @@ namespace Files.Interacts
         {
             var openedPopups = VisualTreeHelper.GetOpenPopups(Window.Current);
             return openedPopups.Any(popup => popup.Child is ContentDialog);
-        }
-
-        /// <summary>
-        /// Navigates to a directory or opens file
-        /// </summary>
-        /// <param name="path"></param>
-        /// <param name="itemType"></param>
-        /// <param name="openSilent">Determines whether history of opened item is saved (... to Recent Items/Windows Timeline/opening in background)</param>
-        /// <param name="openViaApplicationPicker">Determines whether open file using application picker</param>
-        /// <param name="selectItems">List of filenames that are selected upon navigation</param>
-        public async Task<bool> OpenPath(string path, FilesystemItemType? itemType = null, bool openSilent = false, bool openViaApplicationPicker = false, IEnumerable<string> selectItems = null)
-        // TODO: This function reliability has not been extensively tested
-        {
-            string previousDir = AssociatedInstance.FilesystemViewModel.WorkingDirectory;
-            bool isHiddenItem = NativeFileOperationsHelper.HasFileAttribute(path, System.IO.FileAttributes.Hidden);
-            bool isShortcutItem = path.EndsWith(".lnk") || path.EndsWith(".url"); // Determine
-            FilesystemResult opened = (FilesystemResult)false;
-
-            // Shortcut item variables
-            string shortcutTargetPath = null;
-            string shortcutArguments = null;
-            string shortcutWorkingDirectory = null;
-            bool shortcutRunAsAdmin = false;
-            bool shortcutIsFolder = false;
-
-            if (itemType == null || isShortcutItem || isHiddenItem)
-            {
-                if (isShortcutItem)
-                {
-                    var (status, response) = await Connection.SendMessageForResponseAsync(new ValueSet()
-                    {
-                        { "Arguments", "FileOperation" },
-                        { "fileop", "ParseLink" },
-                        { "filepath", path }
-                    });
-
-                    if (status == AppServiceResponseStatus.Success)
-                    {
-                        shortcutTargetPath = response.Get("TargetPath", string.Empty);
-                        shortcutArguments = response.Get("Arguments", string.Empty);
-                        shortcutWorkingDirectory = response.Get("WorkingDirectory", string.Empty);
-                        shortcutRunAsAdmin = response.Get("RunAsAdmin", false);
-                        shortcutIsFolder = response.Get("IsFolder", false);
-
-                        itemType = shortcutIsFolder ? FilesystemItemType.Directory : FilesystemItemType.File;
-                    }
-                    else
-                    {
-                        return false;
-                    }
-                }
-                else if (isHiddenItem)
-                {
-                    itemType = NativeFileOperationsHelper.HasFileAttribute(path, System.IO.FileAttributes.Directory) ? FilesystemItemType.Directory : FilesystemItemType.File;
-                }
-                else
-                {
-                    itemType = await StorageItemHelpers.GetTypeFromPath(path);
-                }
-            }
-
-            var mostRecentlyUsed = Windows.Storage.AccessCache.StorageApplicationPermissions.MostRecentlyUsedList;
-
-            if (itemType == FilesystemItemType.Directory) // OpenDirectory
-            {
-                if (isShortcutItem)
-                {
-                    if (string.IsNullOrEmpty(shortcutTargetPath))
-                    {
-                        await InvokeWin32ComponentAsync(path);
-                        return true;
-                    }
-                    else
-                    {
-                        AssociatedInstance.NavigationToolbar.PathControlDisplayText = shortcutTargetPath;
-                        AssociatedInstance.NavigateWithArguments(AssociatedInstance.InstanceViewModel.FolderSettings.GetLayoutType(shortcutTargetPath), new NavigationArguments()
-                        {
-                            NavPathParam = shortcutTargetPath,
-                            AssociatedTabInstance = AssociatedInstance,
-                            SelectItems = selectItems
-                        });
-
-                        return true;
-                    }
-                }
-                else if (isHiddenItem)
-                {
-                    AssociatedInstance.NavigationToolbar.PathControlDisplayText = path;
-                    AssociatedInstance.NavigateWithArguments(AssociatedInstance.InstanceViewModel.FolderSettings.GetLayoutType(path), new NavigationArguments()
-                    {
-                        NavPathParam = path,
-                        AssociatedTabInstance = AssociatedInstance
-                    });
-
-                    return true;
-                }
-                else
-                {
-                    opened = await AssociatedInstance.FilesystemViewModel.GetFolderWithPathFromPathAsync(path)
-                        .OnSuccess(childFolder =>
-                        {
-                            // Add location to MRU List
-                            mostRecentlyUsed.Add(childFolder.Folder, childFolder.Path);
-                        });
-                    if (!opened)
-                    {
-                        opened = (FilesystemResult)FolderHelpers.CheckFolderAccessWithWin32(path);
-                    }
-                    if (!opened)
-                    {
-                        opened = (FilesystemResult)path.StartsWith("ftp:");
-                    }
-                    if (opened)
-                    {
-                        AssociatedInstance.NavigationToolbar.PathControlDisplayText = path;
-                        AssociatedInstance.NavigateWithArguments(AssociatedInstance.InstanceViewModel.FolderSettings.GetLayoutType(path), new NavigationArguments()
-                        {
-                            NavPathParam = path,
-                            AssociatedTabInstance = AssociatedInstance,
-                            SelectItems = selectItems
-                        });
-                    }
-                }
-            }
-            else if (itemType == FilesystemItemType.File) // OpenFile
-            {
-                if (isShortcutItem)
-                {
-                    if (string.IsNullOrEmpty(shortcutTargetPath))
-                    {
-                        await InvokeWin32ComponentAsync(path);
-                    }
-                    else
-                    {
-                        if (!path.EndsWith(".url"))
-                        {
-                            StorageFileWithPath childFile = await AssociatedInstance.FilesystemViewModel.GetFileWithPathFromPathAsync(shortcutTargetPath);
-                            if (childFile != null)
-                            {
-                                // Add location to MRU List
-                                mostRecentlyUsed.Add(childFile.File, childFile.Path);
-                            }
-                        }
-                        await InvokeWin32ComponentAsync(shortcutTargetPath, shortcutArguments, shortcutRunAsAdmin, shortcutWorkingDirectory);
-                    }
-                    opened = (FilesystemResult)true;
-                }
-                else if (isHiddenItem)
-                {
-                    await InvokeWin32ComponentAsync(path);
-                }
-                else
-                {
-                    opened = await AssociatedInstance.FilesystemViewModel.GetFileWithPathFromPathAsync(path)
-                        .OnSuccess(async childFile =>
-                        {
-                            // Add location to MRU List
-                            mostRecentlyUsed.Add(childFile.File, childFile.Path);
-
-                            if (openViaApplicationPicker)
-                            {
-                                LauncherOptions options = new LauncherOptions
-                                {
-                                    DisplayApplicationPicker = true
-                                };
-                                await Launcher.LaunchFileAsync(childFile.File, options);
-                            }
-                            else
-                            {
-                                //try using launcher first
-                                bool launchSuccess = false;
-
-                                StorageFileQueryResult fileQueryResult = null;
-
-                                //Get folder to create a file query (to pass to apps like Photos, Movies & TV..., needed to scroll through the folder like what Windows Explorer does)
-                                StorageFolder currFolder = await AssociatedInstance.FilesystemViewModel.GetFolderFromPathAsync(Path.GetDirectoryName(path));
-
-                                if (currFolder != null)
-                                {
-                                    QueryOptions queryOptions = new QueryOptions(CommonFileQuery.DefaultQuery, null);
-
-                                    //We can have many sort entries
-                                    SortEntry sortEntry = new SortEntry()
-                                    {
-                                        AscendingOrder = FolderSettings.DirectorySortDirection == Microsoft.Toolkit.Uwp.UI.SortDirection.Ascending
-                                    };
-
-                                    //Basically we tell to the launched app to follow how we sorted the files in the directory.
-
-                                    var sortOption = FolderSettings.DirectorySortOption;
-
-                                    switch (sortOption)
-                                    {
-                                        case Enums.SortOption.Name:
-                                            sortEntry.PropertyName = "System.ItemNameDisplay";
-                                            queryOptions.SortOrder.Clear();
-                                            queryOptions.SortOrder.Add(sortEntry);
-                                            break;
-
-                                        case Enums.SortOption.DateModified:
-                                            sortEntry.PropertyName = "System.DateModified";
-                                            queryOptions.SortOrder.Clear();
-                                            queryOptions.SortOrder.Add(sortEntry);
-                                            break;
-
-                                        case Enums.SortOption.DateCreated:
-                                            sortEntry.PropertyName = "System.DateCreated";
-                                            queryOptions.SortOrder.Clear();
-                                            queryOptions.SortOrder.Add(sortEntry);
-                                            break;
-
-                                        //Unfortunately this is unsupported | Remarks: https://docs.microsoft.com/en-us/uwp/api/windows.storage.search.queryoptions.sortorder?view=winrt-19041
-                                        //case Enums.SortOption.Size:
-
-                                        //sortEntry.PropertyName = "System.TotalFileSize";
-                                        //queryOptions.SortOrder.Clear();
-                                        //queryOptions.SortOrder.Add(sortEntry);
-                                        //break;
-
-                                        //Unfortunately this is unsupported | Remarks: https://docs.microsoft.com/en-us/uwp/api/windows.storage.search.queryoptions.sortorder?view=winrt-19041
-                                        //case Enums.SortOption.FileType:
-
-                                        //sortEntry.PropertyName = "System.FileExtension";
-                                        //queryOptions.SortOrder.Clear();
-                                        //queryOptions.SortOrder.Add(sortEntry);
-                                        //break;
-
-                                        //Handle unsupported
-                                        default:
-                                            //keep the default one in SortOrder IList
-                                            break;
-                                    }
-
-                                    fileQueryResult = currFolder.CreateFileQueryWithOptions(queryOptions);
-
-                                    var options = new LauncherOptions
-                                    {
-                                        NeighboringFilesQuery = fileQueryResult
-                                    };
-
-                                    // Now launch file with options.
-                                    launchSuccess = await Launcher.LaunchFileAsync(childFile.File, options);
-                                }
-
-                                if (!launchSuccess)
-                                {
-                                    await InvokeWin32ComponentAsync(path);
-                                }
-                            }
-                        });
-                }
-            }
-
-            if (opened.ErrorCode == FileSystemStatusCode.NotFound && !openSilent)
-            {
-                await DialogDisplayHelper.ShowDialogAsync("FileNotFoundDialog/Title".GetLocalized(), "FileNotFoundDialog/Text".GetLocalized());
-                AssociatedInstance.NavigationToolbar.CanRefresh = false;
-                await CoreApplication.MainView.CoreWindow.Dispatcher.RunAsync(CoreDispatcherPriority.Normal, () =>
-                {
-                    var ContentOwnedViewModelInstance = AssociatedInstance.FilesystemViewModel;
-                    ContentOwnedViewModelInstance?.RefreshItems(previousDir);
-                });
-            }
-
-            return opened;
-        }
-
-        public async void OpenSelectedItems(bool openViaApplicationPicker = false)
-        {
-            if (AssociatedInstance.FilesystemViewModel.WorkingDirectory.StartsWith(AppSettings.RecycleBinPath))
-            {
-                // Do not open files and folders inside the recycle bin
-                return;
-            }
-            if (AssociatedInstance.SlimContentPage == null)
-            {
-                return;
-            }
-            foreach (ListedItem item in AssociatedInstance.SlimContentPage.SelectedItems)
-            {
-                var type = item.PrimaryItemAttribute == StorageItemTypes.Folder ?
-                    FilesystemItemType.Directory : FilesystemItemType.File;
-                await OpenPath(item.ItemPath, type, false, openViaApplicationPicker);
-            }
         }
 
         public async void LaunchNewWindow()
@@ -647,23 +261,6 @@ namespace Files.Interacts
         {
             item.IsHiddenItem = isHidden;
             AssociatedInstance.SlimContentPage.ResetItemOpacity();
-        }
-
-        public void CopyWorkingLocation()
-        {
-            try
-            {
-                if (AssociatedInstance.SlimContentPage != null)
-                {
-                    DataPackage data = new DataPackage();
-                    data.SetText(AssociatedInstance.FilesystemViewModel.WorkingDirectory);
-                    Clipboard.SetContent(data);
-                    Clipboard.Flush();
-                }
-            }
-            catch
-            {
-            }
         }
 
         public async Task PasteItemAsync(string destinationPath)
